refactoring: component
